Send event when device is paired/unpaired and update mDNS service reachability
The mDNS txt record entry 'sf' indicates the reachability of the transport.
If a device is already paired with the transport, sf should be set to 0.
iOS 9 checks for this flag and does not connect to the accessory if it still
set to 1.

Now when a device is paired or unpaired, an event is send using an event emitter
and the mDNS txt records are updated.

This commit fixes #2.