Fix dnssd on Linux to respond to multicast DNS lookups.
dnssd was not joining the multicast DNS IPs 224.0.0.251 / [ff02::fb]

If an interface had not otherwise joined the multicast IPs,
lookup packets were not being seen.

I believe on macOS it was working despite this because
interfaces have typically already joined 224.0.0.251.

On Linux, not seeing the lookups was causing the record to disappear
after the expiration of the SRV announcement.  Registered HomeKit
devices would then be reported as "not responding".

Fixes #4.