Allow events to send the same event multiple times
The removed lines prevent an event to trigger a new event
when the previous triggered event has the same value.

E.g. When implementing a StatelessProgrammableSwitch, with these lines
in place, you can't trigger a single press twice. You first have to send
a double or long press before a single press can be triggered again.